requirements.txt
asn1crypto==0.24.0
Babel==2.5.3
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
Django==1.11
django-phonenumber-field==2.0.0
django-twilio==0.8.0
enum34==1.1.6
httplib2==0.10.3
idna==2.6
ipaddress==1.0.19
phonenumberslite==8.8.9
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.5.0
pytz==2017.3
requests==2.18.4
six==1.11.0
twilio==4.9.0
urllib3==1.22
